put rust test in separate file
